Make it feasible to implement AnyObj Vtable in C ffi
This introduces `RawStr`, which is roughly equivalent to `&'static str`,
but has a well-defined layout that can be used in C.

We also add safe constant conversion functions from `&'static str` to
`RawStr` in Rust.

This is the last piece of the puzzle needed to make `AnyObjVtable`
implementable outside of Rust, which is important to support a future C
ffi.