Rework how protocol functions are called
This allows them to be called more flexibly, and for the `string_debug`
case avoids the need to allocate a local string buffer internally when
debug formatting.

Unfortunately this means the `fmt::Debug` impl suffers a bit in that it
duplicates `string_debug`. There's just no way to punt a
`fmt::Formatter` into a `String`, and since it has a lifetime it can't
be called through rune interfaces.