Make Formatter functions public
The Formatter is needed for various interactions with a rune values. Therefore I need to be able to get a string as result out of the Formatter.

Why is this needed?
When receiving a generic rune::Value from the vm, this value can be a Result. To debug format the error result in rust, the formatter is now needed. Unfortunately the value can't be received out of the Formatter as of now.